// apps/web/src/types/generated/MarketHistoryRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CivilDate } from "./CivilDate";
import type { StockCode } from "./StockCode";

export type MarketHistoryRequest = {
  code: StockCode;
  date_from: CivilDate;
  date_to: CivilDate;
  after: CivilDate | null;
  page_size: number;
};